## Type of change
Build upgrades & maintenance

## Description of change
Lots of build maintenance relating to webpack, istanbul, and karma.

### Fixed bugs
  - Failed tests now use the correct line numbers in the stack traces
  - Code coverage reports use pre-transpiled code now
  - Output from `gulp test —browserstack` is much more concise

### Maintenance
  - Many dependencies (notably Karma and Webpack) run the latest versions
  - CI now uses Node 7 (for Webpack 3.0 compatibility. We should replace yarn with Node 8 in a separate PR)
  - Added `gulp test-coverage` command. `gulp test` no longer generates coverage. `gulp run-tests` uses `gulp test-coverage`, while `gulp serve` uses `gulp test`.
  - PhantomJS replaced by HeadlessChrome everywhere
  - IE9 removed from the browserstack list (since we don't support it anymore)

## Performance Comparison
These were very un-scientific... I only ran each test once, and didn't pay attention to my machine load. I can run more careful trials if anyone really cares though.

### Old build (from Master)
prebid.js bundle size — 269KB

`npm install` — 31s
`yarn install` — 20.45s
`gulp build` — 13s
`gulp lint` — 6s
`gulp run-tests` — 1m1s

## New build (from this branch)
prebid.js bundle size — 264KB

`npm install` — 25s
`yarn install` — 18.59s
`gulp build` — 15s
`gulp lint` — 5s
`gulp test` — 37s
`gulp run-tests` — 1m17s (no longer used by `gulp serve`)
`gulp test --browserstack` -- About 10 minutes