@mkendall07 2017-07-13
@snapwich appreciate your review on these changes.